YLW-3119-4 organise folder structure.
