+++ TheEnchantedLibrary/Actions/ListBooks.cs
@@ -1,4 +1,7 @@
-﻿namespace TheEnchantedLibrary.Models;
+﻿using TheEnchantedLibrary.Models.Interfaces;
+using TheEnchantedLibrary.Services.Interfaces;
+
+namespace TheEnchantedLibrary.Actions;
 
 public class ListBooks : LibraryAction
 {
