+++ TheEnchantedLibrary/Actions/RemoveBook.cs
@@ -1,4 +1,7 @@
-﻿namespace TheEnchantedLibrary.Models;
+﻿using TheEnchantedLibrary.Models.Interfaces;
+using TheEnchantedLibrary.Services.Interfaces;
+
+namespace TheEnchantedLibrary.Actions;
 
 public class RemoveBook : LibraryAction
 {
@@ -17,12 +20,13 @@ public class RemoveBook : LibraryAction
     {
         _libraryUserInteraction.PrintMessage("Remove book.");
         var title = _libraryUserInteraction.ReadInput("Enter the title of the book to remove: ");
+        Console.WriteLine(title);
         try
         {
             var book = library.GetBooks().First(x => x.Title == title);
             library.RemoveBook(book);
             _libraryUserInteraction.PrintMessage("");
-            _libraryUserInteraction.PrintMessage($"| {title} removed. |");
+            _libraryUserInteraction.PrintMessage($"{title} removed.");
             _libraryUserInteraction.PrintMessage("");
         }
         catch (InvalidOperationException)
