+++ TheEnchantedLibrary/Actions/Search.cs
@@ -1,12 +1,15 @@
-﻿namespace TheEnchantedLibrary.Models;
+﻿using TheEnchantedLibrary.Models.Interfaces;
+using TheEnchantedLibrary.Services.Interfaces;
 
-public class ExitApp : LibraryAction
+namespace TheEnchantedLibrary.Actions;
+
+public class Search : LibraryAction
 {
     public override string Message { get; } = "E[x]it application" + Environment.NewLine;
 
     private readonly ILibraryUserInteraction _libraryUserInteraction;
 
-    public ExitApp(ILibraryUserInteraction libraryUserInteraction)
+    public Search(ILibraryUserInteraction libraryUserInteraction)
     {
 
         _libraryUserInteraction = libraryUserInteraction;
