--- TheEnchantedLibrary/Models/LibraryAction.cs
@@ -1,10 +1,0 @@
-﻿using TheEnchantedLibrary.Models;
-
-namespace TheEnchantedLibrary;
-
-public abstract class LibraryAction : ILibraryAction
-{
-    public abstract string Message { get; }
-
-    public abstract void Execute(ILibrary library);
-}
